requirements.txt
click==8.1.3
marshmallow-dataclass==8.5.9
scikit-learn==1.1.2
pyyaml==5.1
mlflow==1.22.0
pandas==1.5.0
protobuf==3.20.*
pytest-order==1.0.1